requirements.txt
asgiref             
certifi             
charset-normalizer
Django            
djangorestframework 
idna      
matplotlib             
numpy                  
opencv-python          
opencv-python-headless 
paho-mqtt              
pandas                 
pip 
psycopg2            
psycopg2-binary                    
pyparsing              
pyserial               
python-dateutil        
pytz                   
PyYAML                 
requests               
scipy   
setuptools          
sqlparse  
supervision
torch          
torchvision            
ultralytics       
urllib3             
wheel               

